+++ ChangeLog.md
@@ -4,7 +4,6 @@
 
 ## Upcoming Release
 
-
 General:
 
 - Bump mysql2 to resolve to 3.10.1 for security patches
@@ -13,7 +12,7 @@ Blob:
 
 - Fixed issue of download 0 size blob with range > 0 should report error. (issue #2410)
 - Fixed issue of download a blob range without header x-ms-range-get-content-md5, should not return content-md5. (issue #2409)
-
+- Supported x-ms-copy-source-tag-option in copy blob from Uri. (issue #2398)
 
 ## 2024.06 Version 3.31.0